tp9 terminado

+++ instrumentos_api/src/main/java/org/example/dtos/PedidoMensualDto.java
@@ -3,5 +3,5 @@ package org.example.dtos;
 public record PedidoMensualDto(
         Integer anio,
         Integer mes,
-        Double total
+        long total
 ) {}
